🚚 rename: DetailedTicket 위치를 organisms로 이동 #58
