+++ src/common/components/organisms/DetailedTicket.tsx
